Added Tooth Tooth Viewer
Added a new component which shows the currently selected tooth.
This uses a single GLB file which contains all the adult teeth.
Teeth are hidden until a selection is given, then unhidden.

Also did a tidy-up of the mouth viewer code. Split the two components over
two files in a folder.
